requirements.txt
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
defusedxml==0.7.1
idna==3.10
pillow==11.3.0
pycparser==2.22
python-dotenv==1.1.1
requests==2.32.4
tinycss2==1.4.0
urllib3==2.5.0
webencodings==0.5.1
